prepare for release
