+++ packages/babel-plugin/build.js
@@ -2,7 +2,12 @@ const { build } = require('esbuild')
 const { peerDependencies, dependencies } = require('./package.json')
 
 build({
-  entryPoints: ['src/index.ts', 'src/fetch-images.ts'],
+  entryPoints: [
+    'src/index.ts',
+    'src/compiler.ts',
+    'src/config.ts',
+    'src/fetch-images.ts',
+  ],
   outdir: 'dist',
   bundle: true,
   platform: 'node',
