+++ packages/babel-plugin/src/__fixtures__/site.js
@@ -71,7 +71,9 @@ export default function Index() {
         spaceBetween={{ 'breakpoints.large': 'minmax(24px, 0.5fr)' }}
         spaceYStart="40px"
         spaceYEnd="80px"
-        // background={{ xray: 'url(#diagonalHatch)' }}
+        background={{ xray: 'url(#diagonalHatch)' }}
+        // TODO: add X/YWrap components
+        style={{ flexWrap: 'wrap' }}
       >
         {['Elements', 'Variants', 'Overrides', 'Editor'].map((name) => (
           <Stack
